WLANRTSYS-21502:Can't connect STA after adding more than 8 VAPs
At the time of recovery, reboot and factory reset, when deactivation of vap is happening, the bss reserved queue is cleaned once for every vap by looping the queue. As the lock is same for both bss ring and bss reserved queue, cpu stalls are observed at the time of bss reserved queue clean up.
As part of this fix bss reserved queue clean up is done at the time last vap removal of radio instead of cleanup for every vap

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>